requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncmy==0.2.10
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
fastapi==0.115.14
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
Werkzeug==3.1.3
